minion/requirements.txt
python-dotenv==0.15.0
pika==1.1.0
pymongo==3.6.0 
